requirements.txt
altgraph==0.17.2
asciimatics==1.13.0
auto-py-to-exe==2.15.0
bottle==0.12.19
bottle-websocket==0.2.9
cachetools==5.0.0
cffi==1.15.0
Eel==0.12.4
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
pefile==2021.9.3
Pillow==9.0.0
py2exe==0.11.0.1
pycparser==2.21
pyfiglet==0.8.post1
pyinstaller==4.8
pyinstaller-hooks-contrib==2022.0
python-magic-bin==0.4.14
pywin32==303
pywin32-ctypes==0.2.0
Wand==0.6.7
wcwidth==0.2.5
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0
